+++ CHANGELOG.md
@@ -6,6 +6,12 @@ Subheadings to categorize changes are `added, changed, deprecated, removed, fixe
 
 ## Unreleased
 ### added
+- Added Cargo feature directives to select features from the HTML file.
+  - `data-cargo-features`: (optional) Space or comma separated list of Cargo features to activate.
+  - `data-cargo-no-default-features`: (optional) Disables the default Cargo features.
+  - `data-cargo-all-features`: (optional) Enables all Cargo features
+    - Neither compatible with `data-cargo-features` nor `data-cargo-no-default-features`.`data-cargo-features`.
+
 ### changed
 ### fixed